Implemented OW_IGNORE_EGGS_ON_HEAL (#4888)
* Implemented OW_IGNORE_EGGS_ON_HEAL

* Removed videos and change nPokemon to u32

* Changed OW_IGNORE_EGGS_ON_HEAL to GEN_LATEST per https://github.com/rh-hideout/pokeemerald-expansion/pull/4888\#issuecomment-2205271486

* Update include/config/overworld.h

Fixed spelling error per https://github.com/rh-hideout/pokeemerald-expansion/pull/4888#discussion_r1664454436


---------

+++ src/pokemon_storage_system.c
@@ -1436,9 +1436,9 @@ s16 GetFirstFreeBoxSpot(u8 boxId)
     return -1; // all spots are taken
 }
 
-u8 CountPartyNonEggMons(void)
+u32 CountPartyNonEggMons(void)
 {
-    u16 i, count;
+    u32 i, count;
 
     for (i = 0, count = 0; i < PARTY_SIZE; i++)
     {
@@ -6538,7 +6538,7 @@ static void ReleaseMon(void)
             if (OW_PC_RELEASE_ITEM >= GEN_8)
                 item = GetBoxMonDataAt(boxId, sCursorPosition, MON_DATA_HELD_ITEM);
         }
-            
+
         PurgeMonOrBoxMon(boxId, sCursorPosition);
         if (item != ITEM_NONE)
             AddBagItem(item, 1);
